Fix ENTER key in diff mode, add options

+++ wdisk.sh
@@ -10,6 +10,7 @@ ovmf=$(dirname "$0")/../src/OVMF.fd0
 flags=
 interface="if=virtio"
 efi="-pflash $ovmf.tmp"
+acc=hvf
 while [ $# != 0 ]; do
   case "$1" in
 	*.iso|*.dmg)
@@ -26,9 +27,19 @@ while [ $# != 0 ]; do
 	/dev/*)
 	  flags="$flags -drive file=$1,format=raw,$interface"
 	  set -xe
-	  diskutil unmountDisk force $1
+	  if $(type diskutil); then
+		diskutil unmountDisk force $1
+	  else
+		for partition in $(ls $1?*); do
+		  if findmnt $partition; then
+			echo UNMOUNT DISABLED! udisksctl unmount -b $partition
+		  fi
+		done
+	  fi
 	  sudo chown $(whoami) $1
-	  diskutil unmountDisk force $1
+	  if $(type diskutil); then
+		diskutil unmountDisk force $1
+	  fi		
 	  ;;
 	+usb1)
 	  # TODO Only 1 usb stick supported 
@@ -50,6 +61,9 @@ while [ $# != 0 ]; do
 	+sd)
 	  interface="if=ide"
 	  ;;
+	+vi)
+	  interface="if=virtio"
+	  ;;
 	-*)
 	  break
 	  ;;
@@ -60,16 +74,16 @@ while [ $# != 0 ]; do
   shift
 done
 set -xe
-cp $ovmf $ovmf.tmp
+test -z "$efi" || test -f $ovmf.tmp || cp $ovmf $ovmf.tmp
 stty intr "^]"
-nice caffeinate\
+caff=$(which caffeinate) || acc=kvm
+nice $caff\
 	 ${QEMU-~/src/build.qemu6/qemu-system-x86_64}\
 	 -m 2.5G\
-	 -display default,show-cursor=on -accel hvf -smp 2\
+	 -display default,show-cursor=on -accel $acc -smp 2\
 	 $flags\
 	 -nic user,model=virtio-net-pci,hostfwd=tcp::3073-:73\
-	 -serial stdio\
-	 -parallel none\
+	 -serial stdio -parallel none\
 	 -device virtio-tablet-pci\
 	 $efi "$@"
 stty intr "^C"
